Organise the files
